+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/aura_extensions/sulu-extension.js
@@ -187,6 +187,26 @@
             };
 
             /**
+             * deletes data locally and in the database
+             * @param key
+             */
+            app.sandbox.sulu.deleteUserSetting = function(key) {
+                delete app.sandbox.sulu.userSettings[key];
+
+                var data = {
+                    key: key
+                };
+
+                // save to server
+                app.sandbox.util.ajax({
+                    type: 'DELETE',
+                    url: '/admin/security/profile/settings',
+                    data: data
+                });
+            };
+
+
+            /**
              * Shows a standard delete warning dialog
              * @param callback {Function} callback function to execute after dialog got closed. The callback gets always
              *                            executed (with true or false as first argument, whether the dialog got
@@ -410,6 +430,7 @@
 
                         // replace default order by custom order settings
                         gridOptions.url = insertOrderParamsInUrl(gridOptions.url, order);
+                        this.sandbox.emit('sulu.list.preload', gridOptions);
 
                         gridOptions.searchInstanceName = gridOptions.searchInstanceName || toolbarOptions.instanceName;
                         gridOptions.columnOptionsInstanceName =